Move some shared helpers into `workers` directory
These will be needed when we port simulators to workers.
