+++ packages/miniflare/src/plugins/core/index.ts
@@ -16,15 +16,13 @@ import {
   kVoid,
   supportedCompatibilityDate,
 } from "../../runtime";
+import { JsonSchema, Log, MiniflareCoreError, Timers } from "../../shared";
 import {
   Awaitable,
-  JsonSchema,
-  Log,
-  MiniflareCoreError,
-  Timers,
+  CoreBindings,
+  CoreHeaders,
   viewToBuffer,
-} from "../../shared";
-import { CoreBindings, CoreHeaders } from "../../workers";
+} from "../../workers";
 import { getCacheServiceName } from "../cache";
 import { DURABLE_OBJECTS_STORAGE_SERVICE_NAME } from "../do";
 import {
