+++ packages/miniflare/src/plugins/core/modules.ts
@@ -9,12 +9,8 @@ import type estree from "estree";
 import { dim } from "kleur/colors";
 import { z } from "zod";
 import { Worker_Module } from "../../runtime";
-import {
-  MatcherRegExps,
-  MiniflareCoreError,
-  globsToRegExps,
-  testRegExps,
-} from "../../shared";
+import { MiniflareCoreError, globsToRegExps } from "../../shared";
+import { MatcherRegExps, testRegExps } from "../../workers";
 
 const SUGGEST_BUNDLE =
   "If you're trying to import an npm package, you'll need to bundle your Worker first.";
