+++ packages/miniflare/src/shared/index.ts
@@ -1,5 +1,4 @@
 export * from "./timers";
-export * from "./data";
 export * from "./error";
 export * from "./event";
 export * from "./log";
