+++ packages/miniflare/src/workers/index.ts
@@ -1,1 +1,2 @@
 export * from "./core";
+export * from "./shared";
